removing unused lambda and files on EFS filesystem
