+++ archive/deploy-analysis.yml
@@ -57,8 +57,8 @@
     - {src: '{{srcdir}}/archive/utils/', dest: '{{destdir}}/utils/', mode: '754', backup: no, directory_mode: yes }
     - {src: '{{srcdir}}/archive/website/', dest: '{{destdir}}/website/', mode: '754', backup: no, directory_mode: yes }
     - {src: '{{srcdir}}/archive/shwrinfo/', dest: '{{destdir}}/data/shwrinfo/', mode: '754', backup: no, directory_mode: yes }
-    - {src: '{{srcdir}}/archive/website/templates/', dest: '/home/ec2-user/pylibs/templates/', mode: '754', backup: no, directory_mode: yes }
-    - {src: '{{srcdir}}/ukmon_pylib/wmplloc/', dest: '/home/ec2-user/pylibs/python/wmplloc/', mode: '754', backup: no, directory_mode: yes }
+    #- {src: '{{srcdir}}/archive/website/templates/', dest: '/home/ec2-user/pylibs/templates/', mode: '754', backup: no, directory_mode: yes }
+    #- {src: '{{srcdir}}/ukmon_pylib/wmplloc/', dest: '/home/ec2-user/pylibs/python/wmplloc/', mode: '754', backup: no, directory_mode: yes }
 
     - {src: '{{srcdir}}/archive/costs/getCostMetrics.sh', dest: '{{destdir}}/costs/', mode: '754', backup: no }
