--- archive/lambdas/samfunctions/getExtraOrbitFiles/samconfig.toml
@@ -1,11 +1,0 @@
-version = 0.1
-[default]
-[default.deploy]
-[default.deploy.parameters]
-stack_name = "getExtraOrbitFiles"
-s3_bucket = "aws-sam-cli-managed-default-samclisourcebucket-1vz4ek90c2d12"
-s3_prefix = "getExtraOrbitFiles"
-region = "eu-west-2"
-confirm_changeset = true
-capabilities = "CAPABILITY_IAM"
-image_repositories = []
