--- archive/lambdas/samfunctions/getExtraOrbitFiles/template.yaml
@@ -1,54 +1,0 @@
-#
-# arn:aws:lambda:eu-west-2:317976261112:function:getExtraFiles
-# remember to add policy to allow s3 bucket to invoke 
-# then add s3 trigger on the properties page of the bucket
-# 
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: AWS::Serverless-2016-10-31
-Parameters:
-  SecurityGroupIds:
-    Type: CommaDelimitedList
-    Default: sg-0839c64ae44d8115c
-  SubnetIDs:
-    Type: CommaDelimitedList
-    Description: The list of SubnetIDs in your Virtual Private Cloud (VPC)
-    Default: subnet-0642990afd740a30d
-  EFSpath:
-    Type: String
-    Default: /mnt/pylibs
-  AccessPointARN:
-    Type: String
-    Description: Access point ARN
-    Default: arn:aws:elasticfilesystem:eu-west-2:317976261112:access-point/fsap-0dfa51e2ab5dabaff
-Resources:
-  createFiles:
-    Type: AWS::Serverless::Function
-    Properties:
-      FunctionName: getExtraFiles
-      Description: Gets UKMON orbit extra files
-      Handler: getExtraOrbitFiles.lambda_handler
-      Runtime: python3.8
-      Timeout: 180
-      MemorySize: 512
-      Environment:
-        Variables:
-          WEBSITEBUCKET: ukmeteornetworkarchive
-          PYTHONPATH: /mnt/pylibs/python
-          TEMPLATES: /mnt/pylibs/templates
-          MPLCONFIGDIR: /tmp/mpl
-      VpcConfig:
-        SecurityGroupIds: !Ref SecurityGroupIds
-        SubnetIds: !Ref SubnetIDs
-      FileSystemConfigs:
-      - Arn: !Ref AccessPointARN
-        LocalMountPath: !Ref EFSpath
-      Role: arn:aws:iam::317976261112:role/lambda-s3-full-access-role
-    xAcctPermission:
-      Type: AWS::Lambda::Permission
-      Properties: 
-        Action: lambda:InvokeFunction
-        FunctionName: !getRef createFiles.Arn 
-        Principal: s3.amazonaws.com
-        SourceAccount: 822069317839
-        SourceArn: arn:aws:s3:::ukmon-shared
-
